services/consumer/requirements.txt
aiokafka==0.10.0
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.1
anthropic==0.18.1
openai==1.12.0
redis==5.0.1
aioredis==2.0.1
pydantic==2.5.3
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2
prometheus-client==0.19.0
numpy==1.26.3
scikit-learn==1.4.0
transformers==4.37.2
torch==2.1.2
whisper==1.1.10